fixing stuff
