+++ .gitignore
@@ -1,3 +1,3 @@
-  
 node_modules
 yarn.lock
+.DS_Store
