--- .travis.yml
@@ -1,7 +1,0 @@
-language: node_js
-node_js:
-  - '12'
-  - '11'
-
-notifications:
-  email: false
