--- changelog.md
@@ -1,18 +1,0 @@
-0.0.3 / 2019-02-21
-
-=================
-
-  * Merge pull request #2 from christianbender/add_bogo_sort
-
-  * changed according to xo-tool
-  * added bogo sort
-  * package.json
-  * Merge pull request #1 from abranhe/donotdisturb
-  * no more notifications
-  * remove comma (,)
-
-0.0.2 / 2018-10-29
-==================
- 
- * fix install issue && jump to 0.0.2
- * init
