--- index.js
@@ -1,4 +1,0 @@
-module.exports = {
-	sorting: require('./src/sorting'),
-	strings: require('./src/strings')
-};
