+++ readme.md
@@ -1,29 +1,20 @@
 <div align="center">
-	<a href="https://npmjs.com/allalgorithms"><img src="https://cdn.abranhe.com/projects/algorithms/logo.svg" width="30%">
-	</a>
 	<br>
 	<br>
+	<img width="300" alt="Ink" src="https://cdn.abranhe.com/projects/algorithms/logo.svg">
 	<br>
-	<br>
-	The All ▲lgorithms Javascript library
-	<br>
+	<a href="https://js.allalgorithms.com"><code>js.allalgorithms.com</code></a>
 	<br>
 </div>
 
-<p align="center">
-	<a href="https://travis-ci.org/abranhe/allalgorithms-js"><img src="https://img.shields.io/travis/abranhe/allalgorithms-js.svg?logo=travis" /></a>
-	<a href="https://github.com/abranhe/allalgorithms-js/blob/master/license"><img src="https://img.shields.io/github/license/abranhe/allalgorithms-js.svg" /></a>
-	<a href="https://github.com/allalgorithms"><img src="https://cdn.abranhe.com/projects/algorithms/badge.svg"/></a>
-	<a href="https://github.com/abranhe/allalgorithms/allalgorithms-js"><img src="https://img.shields.io/npm/v/allalgorithms.svg"/></a>
-</p>
+> The All ▲lgorithms Javascript Library.
 
-<p align="center">
-	<br>
-	<br>
-	<a href="https://js.allalgorithms.com"><code>js.allalgorithms.com</code></a>
-</p>
+[![Build Status](https://github.com/abranhe/allalgorithms-js/workflows/tests/badge.svg)](https://github.com/abranhe/allalgorithms-js/actions)
+[![](https://img.shields.io/github/license/abranhe/allalgorithms-js.svg)](https://github.com/abranhe/allalgorithms-js/blob/master/license)
+[![](https://cdn.abranhe.com/projects/algorithms/badge.svg)](https://github.com/allalgorithms)
+[![](https://img.shields.io/npm/v/allalgorithms.svg)](https://github.com/abranhe/allalgorithms/allalgorithms-js)
 
-# Why?
+## Why?
 
 - Why not 😂
 - Clean and focused
@@ -51,25 +42,25 @@ console.log(allalgorithms.sorting.bubbleSort(arr));
 // -> [-2, 1, 2, 7, 10, 77]
 ```
 
-# Tree
+## Tree
 
 -  **Sorting**
   - [Bubble Sort](https://js.allalgorithms.com/sorting/bubble-sort)
   - [Merge Sort](https://js.allalgorithms.com/sorting/merge-sort)
 
-# Related
+## Related
 
 - [allalgorithms-python](https://github.com/abranhe/allalgorithms-python): All ▲lgorithms Python library
 - [allalgorithms-java](https://github.com/abranhe/allalgorithms-java): All ▲lgorithms Java library
 
-# Maintainers
+## Maintainers
 
 |[![Carlos Abraham Logo][3]][4]|
 | :--------------------------: |
 | [Carlos Abraham][4]          |
 
 
-# License
+## License
 
 [MIT][5] License © [Carlos Abraham][4]
 
@@ -85,5 +76,4 @@ console.log(allalgorithms.sorting.bubbleSort(arr));
 	<a href="https://github.com/abranhe/algorithms">
 		<img src="https://cdn.abranhe.com/projects/algorithms/logo.svg" width="50px">
 	</a>
-  <br>
 </div>
