+++ src/helpers/index.js
@@ -1,0 +1,5 @@
+const Comparator = require('./comparator');
+
+module.exports = {
+  Comparator,
+};
