+++ src/index.js
@@ -1,0 +1,23 @@
+/* Sorting */
+const { bogosort, bubblesort, mergesort } = require('./sorting');
+/* Strings */
+const {
+  anagrams,
+  capitalize,
+  palindrome,
+  reverse,
+  vowels,
+} = require('./strings');
+
+module.exports = {
+  /* Sorting */
+  bogosort,
+  bubblesort,
+  mergesort,
+  /* Strings */
+  anagrams,
+  capitalize,
+  palindrome,
+  reverse,
+  vowels,
+};
