--- src/sorting.js
@@ -1,5 +1,0 @@
-module.exports = {
-	bogoSort: require('./sorting/bogo-sort'),
-	bubbleSort: require('./sorting/bubble-sort'),
-	mergeSort: require('./sorting/merge-sort')
-};
