+++ src/sorting/index.js
@@ -1,0 +1,9 @@
+const bogosort = require('./bogosort');
+const bubblesort = require('./bubblesort');
+const mergesort = require('./mergesort');
+
+module.exports = {
+  bogosort,
+  bubblesort,
+  mergesort,
+};
