--- src/strings.js
@@ -1,7 +1,0 @@
-module.exports = {
-	capitalize: require('./strings/capitalize'),
-	reverse: require('./strings/reverse'),
-	anagrams: require('./strings/anagrams'),
-	palindrome: require('./strings/palindrome'),
-	vowels: require('./strings/vowels')
-};
