+++ src/strings/anagrams.js
@@ -8,7 +8,7 @@
  * @returns {boolean}
  */
 function anagrams(stringA, stringB) {
-	return cleanString(stringA) === cleanString(stringB);
+  return cleanString(stringA) === cleanString(stringB);
 }
 
 /*
@@ -17,7 +17,12 @@ function anagrams(stringA, stringB) {
  * @returns {string}
  */
 function cleanString(str) {
-	return str.replace(/[^\w]/g, '').toLowerCase().split('').sort().join('');
+  return str
+    .replace(/[^\w]/g, '')
+    .toLowerCase()
+    .split('')
+    .sort()
+    .join('');
 }
 
 module.exports = anagrams;
