+++ src/strings/index.js
@@ -1,0 +1,13 @@
+const anagrams = require('./anagrams');
+const capitalize = require('./capitalize');
+const palindrome = require('./palindrome');
+const reverse = require('./reverse');
+const vowels = require('./vowels');
+
+module.exports = {
+  anagrams,
+  capitalize,
+  palindrome,
+  reverse,
+  vowels,
+};
