+++ src/strings/palindrome.js
@@ -7,9 +7,9 @@
  * @returns {boolean}
  */
 function palindrome(str) {
-	return str.split('').every((c, i) => {
-		return c === str[str.length - 1 - i];
-	});
+  return str.split('').every((c, i) => {
+    return c === str[str.length - 1 - i];
+  });
 }
 
 module.exports = palindrome;
