+++ src/strings/vowels.js
@@ -6,11 +6,14 @@
  * @returns {number}
  */
 function vowels(str) {
-	const checker = ['a', 'e', 'i', 'o', 'u'];
+  const checker = ['a', 'e', 'i', 'o', 'u'];
 
-	return str.toLowerCase().split('').reduce((acc, c) => {
-		return checker.includes(c) ? acc + 1 : acc;
-	}, 0);
+  return str
+    .toLowerCase()
+    .split('')
+    .reduce((acc, c) => {
+      return checker.includes(c) ? acc + 1 : acc;
+    }, 0);
 }
 
 module.exports = vowels;
